tests/wpt/tests/workers/WorkerNavigator_appName.htm

<!DOCTYPE html>
<title> WorkerNavigator appName </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var worker = new Worker("./support/WorkerNavigator.js");
  worker.onmessage = t.step_func_done(function(e) {
    assert_equals(e.data.appName, navigator.appName);
  });
});
</script>
